remove test_without_init to make win happy
